Refactoring routers error handlers

+++ src/main/java/uy/com/sofka/biblioteca/usecases/impl/UseCaseBorrar.java
@@ -6,17 +6,20 @@ import org.springframework.validation.annotation.Validated;
 
 import reactor.core.publisher.Mono;
 import uy.com.sofka.biblioteca.repositories.IRecursoRepository;
-import uy.com.sofka.biblioteca.usecases.DeleteRecurso;
+import uy.com.sofka.biblioteca.usecases.BorrarRecurso;
 
 @Service
 @Validated
-public class UseCaseBorrar implements DeleteRecurso {
+public class UseCaseBorrar implements BorrarRecurso {
 
   @Autowired
   private IRecursoRepository repository;
 
   @Override
   public Mono<Void> apply(String id) {
+    // Validar que el id no sea nulo o vacio
+    if(id == null || id.isEmpty())
+      return Mono.error(new Exception("El id del recurso no puede ser nulo o vacio"));
     return repository.deleteById(id).onErrorMap(e -> new IllegalArgumentException("El Id provisto no es valido"));
   }
